// src/api/models/updateFuelStockHistoryRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */

export interface UpdateFuelStockHistoryRequest {
  /** ID */
  id: string;
  /** 리터 당 단가 */
  pricePerLiter: number;
  /** 주유량(L) */
  volume: number;
}